# dist/swagger_output.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "REST API",
        "description": ""
    },
    "host": "localhost:3000",
    "basePath": "/",
    "schemes": [
        "http"
    ],
    "paths": {
        "/health": {
            "get": {
                "description": "",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/rides": {
            "post": {
                "description": "",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "start_lat": {
                                    "example": "any"
                                },
                                "start_long": {
                                    "example": "any"
                                },
                                "end_lat": {
                                    "example": "any"
                                },
                                "end_long": {
                                    "example": "any"
                                },
                                "rider_name": {
                                    "example": "any"
                                },
                                "driver_name": {
                                    "example": "any"
                                },
                                "driver_vehicle": {
                                    "example": "any"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "get": {
                "description": "",
                "parameters": [
                    {
                        "name": "page",
                        "in": "path",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "limit",
                        "in": "path",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/rides/{id}": {
            "get": {
                "description": "",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    }
}